requirements.txt
absl-py==2.1.0
altgraph==0.17.4
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
CTkMessagebox==2.5
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
flatbuffers==24.3.25
fonttools==4.53.0
idna==3.7
iniconfig==2.0.0
jax==0.4.28
jaxlib==0.4.28
kiwisolver==1.4.5
matplotlib==3.9.0
mediapipe==0.10.14
ml-dtypes==0.4.0
numpy==1.26.4
opencv-contrib-python==4.10.0.82
opencv-python==4.10.0.82
opt-einsum==3.3.0
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
pluggy==1.5.0
protobuf==4.25.3
pycparser==2.22
pycryptodomex==3.20.0
pygame==2.5.2
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
pytube==15.0.0
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzipper==0.3.6
Random-Word==1.0.11
requests==2.32.3
scipy==1.13.1
setuptools==70.0.0
six==1.16.0
sounddevice==0.4.7
urllib3==2.2.1
